changes to the organization of word processing
-The word class is put into a seperate script
-Everything should remain the same, this is mostly for organizational purposes in case the word class is used outside word manager.
