+++ Jam 3 Collective 32/Assets/Scenes/Scripts/WordStruc.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[System.Serializable]
+public class Words
+{
+    public string word;
+    public int ID;
+    public float[] Notes;
+    
+
+    //find an Attribute that lets you select the element from 3 choices
+    public string element;
+    public string stance;
+    public string next;
+
+}  
+
