+++ Jam 3 Collective 32/Assets/Scenes/Scripts/WordStruc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c97c67a142026f34da95bf87ff8791c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
